[Silver II] Title: N과 M (12), Time: 76 ms, Memory: 12400 KB -BaekjoonHub

+++ 백준/Silver/15666. N과 M （12）/N과 M （12）.java
@@ -14,7 +14,7 @@ class Main{
             return;
         }
         for(int i=p;i<sz;i++){
-            nb[id]=(int)arr.get(i);
+            nb[id]=arr.get(i);
             seq(id+1,i);
         }
     }
